+++ examples/typescript/backup/user_backup.json
@@ -1,0 +1,9 @@
+[
+  {
+    "_id": "679a1343c406f835121d0f9d",
+    "name": "John Doe",
+    "email": "john.do@example.com",
+    "age": 31,
+    "__v": 0
+  }
+]
