+++ examples/typescript/package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
-    "mongodb-atlas-sdk": "^1.7.12",
+    "mongodb-atlas-sdk": "^1.7.13",
     "mongoose": "^8.9.5",
     "zod": "^3.24.1"
   },
